chore: Re-add Deno getting started
