+++ src/content/docs/get-started.mdx
@@ -14,6 +14,7 @@ next: false
 frameworks:
   - bun
   - bun-hono
+  - deno
   - nest-js
   - next-js
   - node-js
@@ -54,6 +55,12 @@ import BunHonoRequirements from "@/snippets/get-started/bun-hono/Requirements.md
 import BunHonoStep2 from "@/snippets/get-started/bun-hono/Step2.mdx";
 import BunHonoStep3 from "@/snippets/get-started/bun-hono/Step3.mdx";
 import BunHonoStep4 from "@/snippets/get-started/bun-hono/Step4.mdx";
+
+import DenoStep1 from "@/snippets/get-started/deno/Step1.mdx";
+import DenoRequirements from "@/snippets/get-started/deno/Requirements.mdx";
+import DenoStep2 from "@/snippets/get-started/deno/Step2.mdx";
+import DenoStep3 from "@/snippets/get-started/deno/Step3.mdx";
+import DenoStep4 from "@/snippets/get-started/deno/Step4.mdx";
 
 import NestJsStep1 from "@/snippets/get-started/nest-js/Step1.mdx";
 import NestJsRequirements from "@/snippets/get-started/nest-js/Requirements.mdx";
@@ -112,6 +119,9 @@ This guide will show you how to set up a simple API server protected by Arcjet.
 <LinkButton href="/get-started?f=bun-hono" variant="secondary">
   Bun + Hono
 </LinkButton>
+<LinkButton href="/get-started?f=deno" variant="secondary">
+  Deno
+</LinkButton>
 <LinkButton href="/get-started?f=nest-js" variant="secondary">
   NestJS
 </LinkButton>
@@ -141,6 +151,7 @@ In your project root, run the following command:
 <SlotByFramework client:load>
   <BunStep1 slot="bun" />
   <BunHonoStep1 slot="bun-hono" />
+  <DenoStep1 slot="deno" />
   <NestJsStep1 slot="nest-js" />
   <NextJsStep1 slot="next-js" />
   <NodeJsStep1 slot="node-js" />
@@ -155,6 +166,7 @@ In your project root, run the following command:
 <SlotByFramework client:load>
   <BunRequirements slot="bun" />
   <BunHonoRequirements slot="bun-hono" />
+  <DenoRequirements slot="deno" />
   <NestJsRequirements slot="nest-js" />
   <NextJsRequirements slot="next-js" />
   <NodeJsRequirements slot="node-js" />
@@ -173,6 +185,7 @@ project root.
 <SlotByFramework client:load>
   <BunStep2 slot="bun" />
   <BunHonoStep2 slot="bun-hono" />
+  <DenoStep2 slot="deno" />
   <NestJsStep2 slot="nest-js" />
   <NextJsStep2 slot="next-js" />
   <NodeJsStep2 slot="node-js" />
@@ -190,6 +203,7 @@ limit, and prevent bots from accessing your app.
 <SlotByFramework client:load>
   <BunStep3 slot="bun" />
   <BunHonoStep3 slot="bun-hono" />
+  <DenoStep3 slot="deno" />
   <NestJsStep3 slot="nest-js" />
   <NextJsStep3 slot="next-js" />
   <NodeJsStep3 slot="node-js" />
@@ -202,6 +216,7 @@ limit, and prevent bots from accessing your app.
 <SlotByFramework client:load>
   <BunStep4 slot="bun" />
   <BunHonoStep4 slot="bun-hono" />
+  <DenoStep4 slot="deno" />
   <NestJsStep4 slot="nest-js" />
   <NextJsStep4 slot="next-js" />
   <NodeJsStep4 slot="node-js" />
