+++ src/snippets/get-started/deno/Requirements.mdx
@@ -1,0 +1,5 @@
+- Deno 1.46+ if only using the `Deno.serve` API
+
+OR
+
+- Deno 2.0+ if you are using Deno's `serve` command
