+++ src/snippets/get-started/deno/Step1.mdx
@@ -1,0 +1,12 @@
+import SelectableContent from "@/components/SelectableContent";
+
+{/* prettier-ignore */}
+<SelectableContent client:load syncKey="packageManager" frameworkSwitcher>
+<div slot="deno" slotIdx="1">
+
+```bash
+deno add npm:@arcjet/deno
+```
+
+</div>
+</SelectableContent>
