+++ src/snippets/get-started/deno/Step3.mdx
@@ -1,0 +1,12 @@
+import Step3TS from "./Step3.ts?raw";
+import { Code } from "@astrojs/starlight/components";
+import SelectableContent from "@/components/SelectableContent";
+
+<SelectableContent client:load syncKey="language" frameworkSwitcher>
+  <div slot="TS" slotIdx="1">
+    Create a new file at `index.ts` with the contents:
+
+    <Code code={Step3TS} lang="ts" title="index.ts" />
+
+  </div>
+</SelectableContent>
