+++ src/snippets/get-started/deno/Step3.ts
@@ -1,0 +1,46 @@
+import "jsr:@std/dotenv/load";
+
+import arcjet, { detectBot, shield, tokenBucket } from "npm:@arcjet/deno";
+
+const aj = arcjet({
+  key: Deno.env.get("ARCJET_KEY")!, // Get your site key from https://app.arcjet.com
+  characteristics: ["ip.src"], // Track requests by IP
+  rules: [
+    // Shield protects your app from common attacks e.g. SQL injection
+    shield({ mode: "LIVE" }),
+    // Create a bot detection rule
+    detectBot({
+      mode: "LIVE", // Blocks requests. Use "DRY_RUN" to log only
+      // Block all bots except search engine crawlers. See
+      // https://arcjet.com/bot-list
+      allow: ["CATEGORY:SEARCH_ENGINE"],
+    }),
+    // Create a token bucket rate limit. Other algorithms are supported.
+    tokenBucket({
+      mode: "LIVE",
+      refillRate: 5, // Refill 5 tokens per interval
+      interval: 10, // Refill every 10 seconds
+      capacity: 10, // Bucket capacity of 10 tokens
+    }),
+  ],
+});
+
+Deno.serve(
+  { port: 3000 },
+  aj.handler(async (req) => {
+    const decision = await aj.protect(req, { requested: 5 }); // Deduct 5 tokens from the bucket
+    console.log("Arcjet decision", decision.conclusion);
+
+    if (decision.isDenied()) {
+      if (decision.reason.isRateLimit()) {
+        return new Response("Too many requests", { status: 429 });
+      } else if (decision.reason.isBot()) {
+        return new Response("No bots allowed", { status: 403 });
+      } else {
+        return new Response("Forbidden", { status: 403 });
+      }
+    }
+
+    return new Response("Hello world");
+  }),
+);
